✨ mbk-46: refactor workflow and utility packages
Move utility functions to a new `pkg/utils` package for better 
organization. Introduce string formatting and checksum generation 
functions. Create a new `internal/storage` package with interfaces 
for workflow and template storage. This enhances code modularity 
and prepares for future storage implementations.
