+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `v0.1.7` [2025-03-25] restructure project files
+- `v0.1.6` [2025-03-25] emulate default terminal behavior and environment when executing workflows
+- `v0.1.5` [2025-03-25] support concurrent execution of workflows
+- `v0.1.4` [2025-03-25] storage service upgrade
+- `v0.1.3` [2025-03-25] add storage service to support read-only mode in operations
+- `v0.1.2` [2025-03-25] minor bug fixes
+- `v0.1.1` [2025-03-25] minor bug fixes
 - `v0.1.0`  [2024-11-24] add `info` command for workflow details
 - `v0.0.12` [2024-11-22] update shorthand command to mgr over mig
 - `v0.0.11` [2024-11-21] use default user shell over strictly using bash
