+++ cmd/workflow.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/spf13/cobra"
-	"github.com/tesh254/migraine/utils"
+	"github.com/tesh254/migraine/pkg/utils"
 )
 
 var workflowCmd = &cobra.Command{
