+++ internal/execution/shell.go
@@ -1,4 +1,4 @@
-package run
+package execution
 
 import (
 	"bufio"
